+++ src/popups/StagesPopup/ui/RangeLabel.hpp
@@ -2,12 +2,16 @@
 
 #include <Geode/Geode.hpp>
 
+#include "../../../serialization/profile/index.hpp"
+
 using namespace geode::prelude;
 
 class RangeLabel : public CCNode
 {
 private:
+  Range *m_range;
   CCLabelBMFont *m_label;
+  CCLabelBMFont *m_noteLabel;
   bool enabled = false;
 
   const ccColor3B enabledColor = {99, 224, 110};
@@ -16,9 +20,11 @@ private:
   void updateColor();
 
 public:
-  static RangeLabel *create(int from, int to, bool enabled, const std::string &font = "bigFont.fnt");
+  static RangeLabel *create(
+      Range *range,
+      const std::string &font = "gjFont17.fnt");
 
-  bool init(int from, int to, const std::string &font = "bigFont.fnt");
+  bool init(Range *range, const std::string &font = "gjFont17.fnt");
 
   CCLabelBMFont *getLabel() const { return m_label; }